predictService/Maestrik/requirementes.txt
Django==2.0.1
djangorestframework==3.8.2
scikit-learn==0.21.3
pandas==0.24.2
